feat: organize dlc handler related errors, models and constants
