+++ src/constants/dlc-handler.constants.ts
@@ -1,0 +1,7 @@
+export const DLCHandlers = {
+  KEYPAIR: 'keypair',
+  LEATHER: 'leather',
+  UNISAT_FORDEFI: 'unisat/fordefi',
+  LEDGER: 'ledger',
+  DFNS: 'dfns',
+} as const;
