+++ src/index.ts
@@ -1,7 +1,7 @@
 import { DFNSDLCHandler } from './dlc-handlers/dfns-dlc-handler.js';
+import { KeyPairDLCHandler } from './dlc-handlers/keypair-dlc-handler.js';
 import { LeatherDLCHandler } from './dlc-handlers/leather-dlc-handler.js';
 import { LedgerDLCHandler } from './dlc-handlers/ledger-dlc-handler.js';
-import { PrivateKeyDLCHandler } from './dlc-handlers/private-key-dlc-handler.js';
 import { UnisatFordefiDLCHandler } from './dlc-handlers/unisat-fordefi-dlc-handler.js';
 import { EthereumHandler } from './network-handlers/ethereum-handler.js';
 import { RippleHandler } from './network-handlers/ripple-handler.js';
@@ -10,7 +10,7 @@ import { LedgerXRPHandler } from './network-handlers/xrp-ledger-handler.js';
 import { ProofOfReserveHandler } from './proof-of-reserve-handlers/proof-of-reserve-handler.js';
 
 export {
-  PrivateKeyDLCHandler,
+  KeyPairDLCHandler,
   LedgerDLCHandler,
   LeatherDLCHandler,
   UnisatFordefiDLCHandler,
